requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
fastapi==0.117.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
llvmlite==0.45.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numba==0.62.0
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
snorkel==0.9.9
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.36.0
